Refactor dashboard to country-slug routing
